@@ -1,6 +1,9 @@
 import 'dart:convert';
+import 'dart:io';
 
 import 'package:dostop_v2/src/utils/preferencias_usuario.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import '../utils/utils.dart';
 import 'constantes_provider.dart' as constantes;
 
 import 'package:http/http.dart' as http;
@@ -16,7 +19,7 @@ class LoginProvider {
       'contrasena': password,
     };
     try {
-      final resp = await http.post('${constantes.urlApp}/login.php', body: authData);
+      final resp = await http.post(Uri.parse('${constantes.urlApp}/login.php'), body: authData);
       List decodeResp = json.decode(resp.body);
       decodeResp[0].forEach((String k, dynamic v) => mapResp[k] = v);
     } catch (e) {
@@ -37,57 +40,63 @@ class LoginProvider {
     }
   }
 
-
-  Future<Map<String, dynamic>> registrarTokenFCM(String dispositivo,
-      {String idUsuario, String token}) async {
-    if (_prefs.token == '')
+  Future<Map<String, dynamic>> registrarTokenOS(
+      {String? idUsuario, String? playerID}) async {
+    if (_prefs.playerID == '')
       return {
-        'OK': 0,
+        'statusCode': 0,
         'message':
-            'No se ha podido obtener el token desde FCM, revisar que los servicios de Google esten funcionando '
-                'y que los servicios internos de la aplicación se pueden conectar a FCM'
+            'No se ha podido obtener el token desde One Signal, revisar que los servicios esten funcionando '
+                'y que los servicios internos de la aplicación se pueden conectar a One Signal'
       };
-    Map<String, dynamic> mapResp = Map<String, dynamic>();
-    final authData = {
-      'id': idUsuario ?? _prefs.usuarioLogged,
-      'token': token ?? _prefs.token,
-      'dispositivo': dispositivo,
-      'fecha':
-          DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()).toString()
-    };
     try {
+      final deviceData = await getDeviceData();
+      final infoApp = await PackageInfo.fromPlatform();
+      final authData = {
+        'id': _prefs.usuarioLogged,
+        'token': _prefs.token,
+        'playerID': _prefs.playerID,
+        'deviceID': Platform.isIOS ? '1' : '2',
+        'dateAccess':
+            DateFormat('yyyy-MM-dd HH:mm:ss').format(DateTime.now()).toString(),
+        'versionApp': infoApp.version,
+        'SODevice': deviceData['os'],
+        'brand': deviceData['brand'],
+        'model': deviceData['nameModel'],
+      };
       final resp =
-          await http.post('${constantes.urlApp}/registrar_token_disp.php', body: authData);
-      List decodeResp = json.decode(resp.body);
-      decodeResp[0].forEach((String k, dynamic v) => mapResp[k] = v);
+          await http.post(Uri.parse('${constantes.urlApp}/register_playerID.php'), body: authData);
+      Map<String, dynamic> decodeResp = json.decode(resp.body);
+
+      if(resp.statusCode == 200){
+        _prefs.token = '';
+        _prefs.registeredPlayerID = true;
+      }
+
+      return decodeResp;
     } catch (e) {
-      print(
-          'Ocurrió un error en la llamada al Servicio de LOGIN - REGISTRA TOKEN:\n $e');
       return {
-        'OK': 3,
+        'statusCode': 0,
         'message':
             'Ha ocurrido un error, favor verificar conexión a internet. Es posible que no se reciban notificaciones'
       };
     }
-    // print(mapResp);
-    if (mapResp['estatus'] == '1') {
-      return {'OK': 1, 'message': mapResp['estatus']};
-    } else if (mapResp['estatus'] == '2') {
-      return {'OK': 2, 'message': mapResp['message']};
-    }
-    return {'OK': 3, 'message': 'El servicio no respondió con algun mensaje'};
+
   }
 
   Future<bool> logout() async {
     try {
-      final resp = await http.post('${constantes.urlApp}/elimina_token.php',
-          body: {'token': _prefs.token});
-      Map decodeResp = json.decode(resp.body);
-      //print(decodeResp);
+      final resp = await http.post(Uri.parse('${constantes.urlApp}/elimina_token_os.php'),
+          body: {'playerID': _prefs.playerID});
+      Map? decodeResp = json.decode(resp.body);
+      print(decodeResp);
       if (decodeResp == null) return false;
       if (decodeResp.containsKey('estatus')) {
-        if (decodeResp['estatus'] !='3')
+        if (decodeResp['estatus'] !='3') {
+          _prefs.token = '';
+          _prefs.registeredPlayerID = false;
           return true;
+        }
         else
           return false;
       }
