@@ -14,14 +14,13 @@ class VisitantesFreqProvider {
     validaSesion.verificaSesion();
     try {
       final resp = await http.post(
-          '${constantes.urlApp}/obtener_frecuentes.php',
+          Uri.parse('${constantes.urlApp}/obtener_frecuentes.php'),
           body: {'id_colono': idUsuario, 'tipo_frecuente': '$tipo'});
-      Map decodeResp = json.decode(resp.body);
+      Map? decodeResp = json.decode(resp.body);
       final List<VisitanteFreqModel> visitantes = [];
       if (decodeResp == null) return [];
       if (decodeResp.containsKey('lista_frecuente'))
         decodeResp['lista_frecuente'].forEach((visitante) {
-          // print('$visitante');
           final tempFrecuente = VisitanteFreqModel.fromJson(visitante);
           visitantes.add(tempFrecuente);
         });
@@ -34,15 +33,44 @@ class VisitantesFreqProvider {
     }
   }
 
+  Future<Map<String, dynamic>> archivarQR(String idFrecuente) async {
+    try {
+      final url = Uri.parse('${constantes.ulrApiProd}/visita/');
+
+      var request = http.Request('DELETE', url);
+      request.body = json.encode({'idVisitante': idFrecuente});
+      request.headers.addAll({'Content-Type': 'application/json'});
+      final resp = await request.send();
+
+      if (resp.statusCode != 404) {
+        Map mapResp = json.decode(await resp.stream.bytesToString());
+
+        if (mapResp['statusCode'] == 200) {
+          return {'OK': 1};
+        } else {
+          return {'OK': 2, 'message': mapResp['message']};
+        }
+      }
+
+      return {'OK': 2};
+    } catch (e) {
+      print(
+          'Ocurrió un error en la llamada al Servicio de VISITANTES FRECUENTES - ELIMINAR VISITANTE:\n $e');
+
+      return {'OK': 2};
+    }
+  }
+
   Future<Map<String, dynamic>> eliminaVisitanteFrecuente(
       String idFrecuente, String idUsuario, int tipo) async {
     try {
-      final resp = await http
-          .post('${constantes.urlApp}/eliminar_frecuente.php', body: {
-        'id_frecuente': idFrecuente,
-        'id_colono': idUsuario,
-        'tipo_frecuente': '$tipo'
-      });
+      final resp = await http.post(
+          Uri.parse('${constantes.urlApp}/eliminar_frecuente.php'),
+          body: {
+            'id_frecuente': idFrecuente,
+            'id_colono': idUsuario,
+            'tipo_frecuente': '$tipo'
+          });
       Map mapResp = json.decode(resp.body);
       // print(decodeResp);
       if (mapResp['estatus'].toString().contains('1')) {
@@ -57,89 +85,55 @@ class VisitantesFreqProvider {
     }
   }
 
-  Future<Map<String, dynamic>> archivarQR(String idFrecuente) async {
-    try {
-      final url = Uri.tryParse('${constantes.ulrApiProd}/visita/');
-      var request = http.Request('DELETE', url);
-      request.body = json.encode({'idVisitante': idFrecuente});
-      request.headers.addAll({'Content-Type': 'application/json'});
-
-      final resp = await request.send();
-
-      if (resp.statusCode != 404) {
-        Map mapResp = json.decode(await resp.stream.bytesToString());
-        // print(decodeResp);
-        if (mapResp['statusCode'] == 200) {
-          return {'OK': 1};
-        } else {
-          return {'OK': 2, 'message': mapResp['message']};
-        }
-      }
-
-      return {'OK': 2};
-    } catch (e) {
-      print(
-          'Ocurrió un error en la llamada al Servicio de VISITANTES FRECUENTES - ELIMINAR VISITANTE:\n $e');
-      return {'OK': 2};
-    }
-  }
-
   Future<Map<String, dynamic>> nuevoVisitanteFrecuente(
-      {String idUsuario,
-      String nombre,
-      String apPaterno,
-      String apMaterno,
-      String vigencia,
-      String tipoVisitante,
-      String flagOrigen,
-      String telefono}) async {
+      {required String idUsuario,
+      required String nombre,
+      required String apPaterno,
+      required String apMaterno,
+      required String vigencia,
+      required bool esUnico,
+      required String tipoVisitante}) async {
+    Map<String, dynamic> mapResp = Map<String, dynamic>();
     final visitanteData = {
-      'idColonos': idUsuario,
+      'colono': idUsuario,
       'nombre': nombre,
       'ape_paterno': apPaterno,
       'ape_materno': apMaterno,
-      'idTipoVisitante': tipoVisitante,
-      'tipo': flagOrigen == 'dostop' ? 'unico' : '',
-      'vigencia': vigencia == 'indefinido'
-          ? ''
-          : (flagOrigen != 'parco' ? '24 hours' : vigencia),
-      'flagPerpetual': vigencia == 'indefinido' ? true : false,
-      'flagOrigen': flagOrigen,
-      'telefono': flagOrigen == 'parco' ? telefono : '',
-      'usosParco': -1
+      'tipo': esUnico ? 'unico' : '',
+      'vigencia': vigencia,
+      'tipo_visitante': tipoVisitante
     };
     try {
-      final resp = await http.post('${constantes.ulrApiProd}/visita/nueva/',
-          headers: {'Content-Type': 'application/json'},
-          body: json.encode(visitanteData));
-
-      if (resp.statusCode == 404) {
-        return {'statusCode': resp.statusCode, 'status': resp.reasonPhrase};
+      final resp = await http.post(
+          Uri.parse('${constantes.urlApp}/registrar_frecuente2.php'),
+          body: visitanteData);
+      List decodeResp = json.decode(resp.body);
+      decodeResp[0].forEach((k, v) => mapResp[k] = v);
+      // print(decodeResp);
+      if (mapResp['estatus'].toString().contains('1')) {
+        return {
+          'OK': 1,
+          'message': 'Visitante frecuente creado',
+          'codigo': mapResp['codigo']
+        };
+      } else {
+        return {'OK': 2, 'message': mapResp['message']};
       }
-
-      Map<String, dynamic> decodeResp = json.decode(resp.body);
-
-      if (decodeResp.containsKey('statusCode') && resp.statusCode != 200) {
-        return {'statusCode': resp.statusCode, 'status': decodeResp['status']};
-      }
-
-      return decodeResp;
     } catch (e) {
-      var message = e.runtimeType.toString() == 'SocketException'
-          ? 'Ha ocurrido un error, favor verificar conexión a internet.'
-          : e.message;
-      return {'statusCode': 0, 'status': e.runtimeType, 'codigo': message};
+      print(
+          'Ocurrió un error en la llamada al Servicio de VISITANTES FRECUENTES - NUEVO VISITANTE:\n $e');
+      return {'OK': 2};
     }
   }
 
   Future<Map<String, dynamic>> nuevoAccesoRostro(
-      {String idUsuario,
-      String nombre,
-      String apPaterno,
-      String apMaterno,
-      String imgRostroB64,
-      String tipoAcceso,
-      int tipo,
+      {required String idUsuario,
+      required String nombre,
+      required String apPaterno,
+      required String apMaterno,
+      required String imgRostroB64,
+      String? tipoAcceso,
+      required int tipo,
       String tipoVisitante = ""}) async {
     final visitanteData = {
       'id_colono': idUsuario,
@@ -152,7 +146,8 @@ class VisitantesFreqProvider {
       'tipo_visitante': tipoVisitante.toString()
     };
     try {
-      final resp = await http.post('${constantes.urlApp}/registrar_rostro.php',
+      final resp = await http.post(
+          Uri.parse('${constantes.urlApp}/registrar_rostro.php'),
           body: visitanteData);
       Map mapResp = json.decode(resp.body);
       if (mapResp['estatus'].toString().contains('1')) {
@@ -175,19 +170,22 @@ class VisitantesFreqProvider {
   }
 
   Future<Map<String, dynamic>> changeImage(
-      {String idUsuario,
-      String idFrecuente,
-      String image,
-      String tipo}) async {
+      {required String idUsuario,
+      required String idFrecuente,
+      required String image,
+      required String tipo}) async {
     final visitanteData = {
       'idColono': idUsuario,
       'idFrecuente': idFrecuente,
       'image': image,
       'tipoFrecuente': tipo.toString(),
     };
+
     try {
-      final resp = await http.post('${constantes.urlApp}/changeFaceImage.php',
-          body: json.encode(visitanteData), headers: {'Content-Type': 'application/json'});
+      final resp = await http.post(
+          Uri.parse('${constantes.urlApp}/changeFaceImage.php'),
+          body: json.encode(visitanteData),
+          headers: {'Content-Type': 'application/json'});
 
       if (resp.statusCode == 404) {
         return {
@@ -198,10 +196,12 @@ class VisitantesFreqProvider {
       }
 
       Map<String, dynamic> mapResp = json.decode(resp.body);
+
       return mapResp;
     } catch (e) {
       print(
           'Ocurrió un error en la llamada al Servicio de VISITANTES FRECUENTES - NUEVO VISITANTE:\n $e');
+
       return {
         'status': e.runtimeType.toString(),
         'statusCode': 0,
