@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-showOptionPhoto(
-    BuildContext context, Function onPressedTake, Function onPressedGalery) {
+showOptionPhoto(BuildContext context, Function() onPressedTake,
+    Function() onPressedGalery) {
   showCupertinoModalPopup(
       context: context,
       builder: (_) => CupertinoActionSheet(
